requirements.txt
langchain-huggingface
python-dotenv
youtube-transcript-api
langchain-community
langchain-text-splitters
langchain
streamlit
faiss-cpu
sentence-transformers
